ANDX instruction implemented. Also some test code to test the instruction.
Test code was added to main to test out the general functionality of the CPU architecture. Dummy values were used for RAM which was just an array of 32 bit numbers.

+++ src/Bus.cpp
@@ -12,13 +12,13 @@ Bus::~Bus()
 {
 }
 
-void Bus::write(uint16_t addr, uint8_t data)
+void Bus::write(uint16_t addr, uint32_t data)
 {
 	if (addr >= 0x0000 && addr <= 0xFFFF)
 		ram[addr] = data;
 }
 
-uint8_t Bus::read(uint16_t addr, bool bReadOnly)
+uint32_t Bus::read(uint16_t addr, bool bReadOnly)
 {
 	if (addr >= 0x0000 && addr <= 0xFFFF)
 		return ram[addr];
